requirements.txt
flask==2.3.3
gunicorn==21.2.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic[email]==2.9.2
python-dotenv==1.0.1
bcrypt==4.2.0
pytest==8.3.3
PyJWT==2.9.0
flasgger==0.9.7.1